third_party/blink/web_tests/media/media-fragments/TC0017-expected.txt


Title: TC0017
Fragment: 't=,15'
Comment: Equivalent to #t=0,e. The media is requested from 0 to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
